- fix easy signedness problems - mark hard signedness problems as such and remove their warning
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@1994 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/editor/ui_menus/editor_tool_place_bob_options_menu.cc
@@ -134,7 +134,8 @@ void Editor_Tool_Place_Bob_Options_Menu::clicked(int n, bool t) {
 	if (not multiselect) {
 		for (uint i = 0; m_pit.get_nr_enabled(); ++i) m_pit.enable(i, false);
       // Disable all checkboxes
-		for (uint i = 0; i < m_checkboxes.size(); ++i, i += i == n) {
+      // TODO: the uint cast is ugly!
+		for (uint i = 0; i < m_checkboxes.size(); ++i, i += i == (uint)n) {
          m_checkboxes[i]->changedtoid.set(this, &Editor_Tool_Place_Bob_Options_Menu::do_nothing);
          m_checkboxes[i]->set_state(false);
          m_checkboxes[i]->changedtoid.set(this, &Editor_Tool_Place_Bob_Options_Menu::clicked);
